# Implement event timing using cudaEventElapsed