Old:

```
Avg request latency: 3160 usec (overhead 48 usec + queue 127 usec + compute input 222 usec + compute infer 2705 usec + compute output 56 usec)
```

New:

```
 Avg request latency: 3205 usec (overhead 59 usec + queue 153 usec + compute input 162 usec + compute infer 2665 usec + compute output 165 usec)
```

The actual copies for the model take ~100us for both inputs and outputs. I think the new measurements are closer to what is observed in the nsight trace for this model.